% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalreg.R
\name{equalreg}
\alias{equalreg}
\title{Testing the equality of k regression functions}
\usage{
equalreg(
  x,
  y,
  f,
  m_bandwidths = "cv",
  sigma_bandwidths = "cv",
  hmin = 0,
  hmax = 1,
  nh = 30,
  nboot = 500,
  tstat = "KS",
  parallel = FALSE,
  ncores = NULL,
  seed = NULL
)
}
\arguments{
\item{x}{Independent variable.}

\item{y}{Response variable.}

\item{f}{Categorical variable indicating the population to which
the observations belong.}

\item{m_bandwidths}{Bandwidth smoothing parameters for the estimation of the
regression function for each population. Default is "cv" (cross-validation).}

\item{sigma_bandwidths}{Bandwidth smoothing parameters for
the  estimation of the conditional variance function for each population.
Default is "cv".}

\item{hmin}{Minimum value for the bandwidths. Defaults to 0.}

\item{hmax}{Maximum value for the bandwidths. Defaults to 1.}

\item{nh}{Number of points in the grid for the bandwidths. Defaults to 30.}

\item{nboot}{Number of bootstrap repeats. Defaults to 500.}

\item{tstat}{Test statistic to be used. Options include "KS"
(Kolmogorov-Smirnov) or "CM" (Cramér-von Mises) type. Default is "KS".}

\item{parallel}{A logical value. If  \code{TRUE}, the
testing procedure is  parallelized. Note that there are cases
(e.g., a low number of bootstrap repetitions) that R will gain in
performance through serial computation. R takes time to distribute tasks
across the processors also it will need time for binding them all together
later on. Therefore, if the time for distributing and gathering pieces
together is greater than the time need for single-thread computing, it does
not worth parallelize. Defaults to FALSE.}

\item{ncores}{An integer value specifying the number of cores to be used
in the parallelized procedure. If \code{NULL} (default), the number of cores
to be used is equal to the number of cores of the machine - 1.}

\item{seed}{Seed to be used in the procedure. If \code{NULL} (default), the seed is not set.}
}
\value{
A list containing test results,
including the test statistics and p-values.
}
\description{
This function tests the hypothesis of equality of k regression functions
using a bootstrap-based approach.  The test is based on the comparison of
two estimators of the distribution of the errors in each population.
}
\examples{


set.seed(300716)
n1 <- 100
n2 <- 200
x1 <- runif(n1)
y1 <- x1 + 0.25 * rnorm(n1)
x2 <- runif(n2)
y2 <- x2 + 0.25 * rnorm(n2)

x <- c(x1, x2)
y <- c(y1, y2)
f <- c(rep(1, n1), rep(2, n2))
out <- equalreg(x = x, y = y, f = f, tstat = "KS", hmin = 0.05, hmax = 0.5,
nh = 46, nboot = 100, parallel = FALSE, seed = 300716)
out

\dontrun{
library(npregfast)
out2 <- equalreg(x = barnacle$RC, y = barnacle$DW, f = barnacle$F,
tstat = "KS", hmin = 1, hmax = 5, nboot = 2, nh = 15)
}

}
\author{
Marta Sestelo, Nora M. Villanueva and Juan Carlos Pardo-Fernández.
}
