% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.equalreg.R
\name{summary.equalreg}
\alias{summary.equalreg}
\title{Summarizing fits of \code{equalreg} class produced by \code{equalreg}}
\usage{
\method{summary}{equalreg}(object, digits = 4, ...)
}
\arguments{
\item{object}{a equalreg object as producted by \code{equalreg}}

\item{digits}{number of digits to print.}

\item{\ldots}{additional arguments.}
}
\value{
\code{summary.equalreg} computes and returns a list of summary
information for an \code{equalreg} object.

\item{nboot}{Number of bootstrap repeats.}
\item{n}{Total sample size.}
\item{nj}{Sample size of each group.}
\item{test.statistic}{Test statistics used.}
\item{p-values}{p-values of the test.}
\item{components}{A collection of components of the equalreg object.}
}
\description{
Takes a  equalreg object
and produces various useful summaries from it.
}
\details{
\code{print.equalreg} tries to be smart about \code{summary.equalreg}.
}
\examples{

set.seed(300716)
n1 <- 100
n2 <- 200
x1 <- runif(n1)
y1 <- x1 + 0.25 * rnorm(n1)
x2 <- runif(n2)
y2 <- x2 + 0.25 * rnorm(n2)

x <- c(x1, x2)
y <- c(y1, y2)
f <- c(rep(1, n1), rep(2, n2))
out <- equalreg(x = x, y = y, f = f, tstat = "KS", hmin = 0.05, hmax = 0.5,
nh = 46, nboot = 100, cluster = FALSE, seed = 300716)
summary(out)



}
\author{
Marta Sestelo, Nora M. Villanueva and Juan Carlos Pardo-Fernández
}
