% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.equalreg.R
\name{autoplot.equalreg}
\alias{autoplot.equalreg}
\title{Visualization of \code{equalreg} objects with ggplot2 graphics}
\usage{
\method{autoplot}{equalreg}(object = object, interactive = FALSE, type = "all", ...)
}
\arguments{
\item{object}{Object of \code{equalreg} class.}

\item{interactive}{Logical flag indicating if an interactive plot with
plotly is produced.}

\item{type}{Character string indicating the type of plot to be produced. Options
include "mean" (mean regression function), "var" (conditional variances function),
"eps" (empirical distributions functions) or "all" (it includes all) type. Default is "all".}

\item{\ldots}{Other options.}
}
\value{
A ggplot object, so you can use common features from
ggplot2 package to manipulate the plot.
}
\description{
Useful for drawing the estimated functions (regression and
conditional variance functions) for each group or population. Additionally, the empirical
cumulative distribution functions for the residuals of each group are displayed.
}
\examples{

set.seed(300716)
n1 <- 100
n2 <- 200
x1 <- runif(n1)
y1 <- x1 + 0.25 * rnorm(n1)
x2 <- runif(n2)
y2 <- x2 + 0.25 * rnorm(n2)

x <- c(x1, x2)
y <- c(y1, y2)
f <- c(rep(1, n1), rep(2, n2))
out <- equalreg(x = x, y = y, f = f, tstat = "KS", hmin = 0.05, hmax = 0.5,
nh = 46, nboot = 100, parallel = FALSE, seed = 300716)
autoplot(out)
autoplot(out, interactive = TRUE)


}
\author{
Marta Sestelo, Nora M. Villanueva and Juan Carlos Pardo-Fernández.
}
